Small updates to get this working better.
Updates paper_dir to use the tilde for home directory, changes
lit_add, lit_update, mendeley_to_org so they use os.path.expanduser to
reflect this.

Realized that "python script.py" doesn't use the Anaconda version of
python from inside emacs, so changed that python explicitly to the
Anaconda version. I don't know of a way to have emacs check if things
are working, so that will just need to be manually set by the
user. Added something to README.org to reflect this.

Also found a way to have helm-bibtex use the system default
application to open pdfs instead of relying on emacs.