No longer requires all files to have same name
Previously, we were assuming that all files in a given directory had the
same name (and only differed by their extension). This mattered because
of how we updated the paths in the master file (we were assuming the
file name, minus the extension, was the same as the directory the file
was in). This isn't always the case, so now we use the bib_id as the
directory instead.