Starts rework of package into emacs-only
This package renames the literature.el as literature-config.el and
starts a new file, literature.el, which contains the actual code.

literature.el is still a work in progress. At this point,
literature-add-pdf, literature-setup-files,
literature-add-to-master-bib, and literature-add-to-master-org should
all work, as well as a couple helper functions. git-update-commit needs
some more changes, and literature-add-bib hasn't been written at all