Several bugfixes, finalizes literature-add-file
With these bugfixes, literature-add-file should work completely. It is
currently rather slow (largely because of moving to the master org and
bib files and then sorting them), so ideally we can get around
that.

literature-setup-files should now work with an arbitrary file passed as
filepath instead of assuming it must be a pdf

the get-buffer / kill-buffer commands in master bib and org add were
using the wrong variable to get the buffer name. That has been fixed.